Moved utils into setup
