+++ Setup/Parse_Arguments.py
@@ -2,7 +2,7 @@
 import argparse
 from .Setup import Setup
 
-class Parse_Arguments(Setup):
+class parse_arguments(Setup):
     
     def __init__(self):
         super().__init__()
